feat(publication): move processing to a dedicated dir
